feat: modified search algorithm
