+++ src/bgrep.c
@@ -1,5 +1,5 @@
 /* a fork of bgrep
- * m0zikit0 <three1518@163.com>
+ * mozkito <three1518@163.com>
  *
  * Copyright 2009 Felix Domke <tmbinc@elitedvb.net>. All rights reserved.
  *
@@ -36,10 +36,63 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include <fcntl.h>
-#include <unistd.h>
+#define BGREP_VERSION "0.3"
 
-#define BGREP_VERSION "0.2"
+typedef unsigned char uc;
+typedef unsigned long long ull;
+
+/* Sunday algorithm
+ * modified for masked pattern
+ */
+void search(uc *bytes, int size, uc *pattern, uc *mask, int len) {
+  if (size < len) {
+    return;
+  }
+  int i, j, delta[256];
+  uc c, M, m;
+  for (c = 0; c < 256; c++) {
+    delta[c] = len + 1;
+  }
+  // preprocessing
+  for (j = 0; j < len; j++) {
+    if (mask[j] == 0xff) {
+      delta[pattern[j]] = len - j;
+    } else {
+      M = ~mask[j], m = 1;
+      while ((m | M) != M) {
+        m <<= 1;
+      }
+      for (c = 0; c <= M; c += m) {
+        if ((c | M) == M) {
+          delta[c | pattern[j]] = len - j;
+        }
+      }
+    }
+  }
+  // grouped by 8 bytes
+  i = 0;
+  while (size - i >= len) {
+    j = 0;
+    while (j < len) {
+      if (j + 8 > len) {
+        if ((*(ull *)(bytes + i + j) & *(ull *)(mask + j)) << (8 - len + j) !=
+            (*(ull *)(pattern + j)) << (8 - len + j)) {
+          break;
+        }
+      } else if ((*(ull *)(bytes + i + j) & *(ull *)(mask + j)) !=
+                 *(ull *)(pattern + j)) {
+        break;
+      }
+      j += 8;
+    }
+    if (j >= len) {
+      printf("[+] %02X\n", i);
+    }
+    if (i + len < size) {
+      i += delta[bytes[i + len]];
+    }
+  }
+}
 
 void die(const char *msg, ...) {
   va_list ap;
@@ -50,213 +103,45 @@ void die(const char *msg, ...) {
   exit(1);
 }
 
-void print_char(unsigned char c) {
-  if (32 <= c && c <= 126) {
-    putchar(c);
-  } else {
-    printf("\\x%02x", (int)c);
-  }
-}
-
-int ascii2hex(unsigned char c) {
-  if (c >= 0x30 && c <= 0x39) {
-    return c - 0x30;
-  } else if (c >= 0x41 && c <= 0x46) {
-    return c - 0x41 + 10;
-  } else if (c >= 0x61 && c <= 0x66) {
-    return c - 0x61 + 10;
-  } else {
-    return -1;
-  }
-}
-
-int g_bytes_before;
-int g_bytes_after;
-const unsigned char *g_hex;
-const char *g_path;
-int g_pid;
-
-/* TODO: this will not work with STDIN or pipes
- * We have to maintain a window of the bytes before
- * which I am too lazy to do right now.
- */
-void dump_content(int fd, unsigned long long pos) {
-  int save_pos = lseek(fd, 0, SEEK_CUR);
-
-  if (save_pos == -1) {
-    perror("[-] lseek");
-    return; /* this one is not fatal*/
-  }
-
-  char buf[1024];
-  int start = pos - g_bytes_before;
-  int bytes_to_read = g_bytes_before + g_bytes_after;
-
-  if (lseek(fd, start, SEEK_SET) == -1) {
-    perror("[-] lseek");
-    return;
-  }
-
-  while (bytes_to_read > 0) {
-    int read_chunk = bytes_to_read > sizeof(buf) ? sizeof(buf) : bytes_to_read;
-    int bytes_read = read(fd, buf, read_chunk);
-
-    if (bytes_read < 0) {
-      die("[-] error reading context");
-    }
-
-    char *buf_end = buf + bytes_read;
-    char *p = buf;
-
-    for (; p < buf_end; p++) {
-      print_char(*p);
-    }
-
-    bytes_to_read -= read_chunk;
-  }
-
-  putchar('\n');
-
-  if (lseek(fd, save_pos, SEEK_SET) == (int)-1) {
-    die("[-] could not restore the original file offset while printing "
-        "context");
-  }
-}
-
-void search(int fd, const unsigned char *buffer, int size,
-            const unsigned char *pattern, const unsigned char *mask, int len) {
-  int i, j, any_wildcard;
-  for (j = 0, any_wildcard = 0; j < len; j++) {
-    if (mask[j] != 0xFF) {
-      any_wildcard = 1;
-      break;
-    }
-  }
-
-  if (any_wildcard) {
-    printf("[+] searching through brute force\n");
-    i = j = 0;
-    while (size - i >= len) {
-      for (j = 0; j < len; j++) {
-        if ((buffer[i + j] & mask[j]) != pattern[j]) {
-          break;
-        }
-      }
-      if (j == len) {
-        printf(" [+] %s: %08llx\n", g_path, i);
-        if (g_bytes_before || g_bytes_after) {
-          dump_content(fd, i);
-        }
-      }
-      i++;
-    }
-  } else {
-    int k = 0;
-    int *lps = (int *)malloc(sizeof(int) * len);
-    if (lps == NULL)
-      perror("[-] malloc");
-
-    printf("[+] searching through KMP\n");
-    /* pre-process lps array */
-    lps[0] = 0, j = 1;
-    while (j < len) {
-      if (pattern[j] == pattern[k]) {
-        lps[j++] = ++k; /* k pins the longest prefix */
-      } else if (k != 0) {
-        k = lps[k - 1]; /* shorter prefix possible... */
-      } else {
-        lps[j++] = 0; /* no prefix suffix here */
-      }
-    }
-
-    /*
-     * implementation of the KMP algorithm
-     * pre-process the longest prefix suffix value
-     * **produces wrong answers when there's a wild card**
-     */
-    i = j = 0;
-    while (size - i >= len - j) {
-      if (j == len) {
-        /* found a match, continue at the longest prefix suffix */
-        printf(" [+] %s: %08llx\n", g_path, i - j);
-        if (g_bytes_before || g_bytes_after) {
-          dump_content(fd, i - j);
-        }
-        j = lps[j - 1];
-        continue;
-      }
-      if (pattern[j] == buffer[i]) {
-        i++, j++;
-      } else if (j != 0) {
-        j = lps[j - 1]; /* shorter prefix possible */
-      } else {
-        i++; /* give up because no prefix suffix available */
-      }
-    }
-  }
-}
-
-void search_file(int fd, const unsigned char *value, const unsigned char *mask,
-                 int len) {
-  int r, tail = len - 1, offset = 0;
-  size_t bufsize = 1024;
-  /* use a search buffer which is at least the next power of two after len */
-  while (bufsize <= (size_t)len)
-    bufsize <<= 1;
-
-  unsigned char *buf = malloc(bufsize);
-  if (!buf) {
-    die("[-] error allocating search buffer!");
-  }
-
-  while (1) {
-    memmove(buf, buf + bufsize - tail, tail);
-    r = read(fd, buf + tail, bufsize - tail);
-
-    if (r < 0) {
-      perror("read");
-      break;
-    } else if (!r)
-      break;
-
-    search(fd, buf + offset, r, value, mask, len);
-    offset += r;
-  }
-
-  free(buf);
-}
-
 void usage(char **argv) {
   fprintf(stderr, "bgrep version: %s\n", BGREP_VERSION);
-  fprintf(stderr,
-          "usage: %s [-a bytes] [-b bytes] [-c bytes] [-p pid] <hex> <path>\n",
-          *argv);
-  fprintf(stderr, " -a bytes to show after the match\n"
-                  " -b bytes to show before the match\n"
-                  " -c bytes count, both before and after\n"
-                  " -p process id/-f file path\n");
+  fprintf(
+      stderr,
+      "usage: %s [-a bytes] [-b bytes] [-c bytes] [-p pid]|[-f file] <hex>\n",
+      *argv);
+  fprintf(stderr, " -a --bytes-after  bytes to show after the match\n"
+                  " -b --bytes-before bytes to show before the match\n"
+                  " -c --bytes-count  bytes to show before and after\n"
+                  " -p --pid          process id\n"
+                  " -f --file         path of file to read\n");
   exit(1);
 }
 
 typedef enum {
   PARSE_RST,
-  PARSE_HEX,
   PARSE_BAFTER,
   PARSE_BBEFORE,
   PARSE_BCOUNT,
   PARSE_PID,
-  PARSE_PATH
+  PARSE_PATH,
+  PARSE_HEX
 } parse_stat;
 
 struct {
   const char *repr;
   parse_stat opt;
-} g_options[] = {{"-a", PARSE_BAFTER},  {"--bytes-after", PARSE_BAFTER},
+} g_options[] = {{"-p", PARSE_PID},     {"--pid", PARSE_PID},
+                 {"-f", PARSE_PATH},    {"--file", PARSE_PATH},
+                 {"-a", PARSE_BAFTER},  {"--bytes-after", PARSE_BAFTER},
                  {"-b", PARSE_BBEFORE}, {"--bytes-before", PARSE_BBEFORE},
                  {"-c", PARSE_BCOUNT},  {"--bytes-count", PARSE_BCOUNT},
-                 {"-p", PARSE_PID},     {"--pid", PARSE_PID},
-                 {"-f", PARSE_PATH},    {"--file", PARSE_PATH},
                  {NULL, (parse_stat)0}};
+
+uc *g_hex;
+int g_pid;
+char *g_path;
+int g_bytes_after;
+int g_bytes_before;
 
 void parse_opts(int argc, char **argv) {
   int i = 1, j, k;
@@ -330,15 +215,6 @@ void parse_opts(int argc, char **argv) {
     die("[-] invalid value %d for pid", g_pid);
   }
 }
-
-/* The Windows/DOS implementation of read(3) opens files in text mode by
- * default, which means that an 0x1A byte is considered the end of the file
- * unless a non-standard flag is used. Make sure it's defined even on real POSIX
- * environments
- */
-#ifndef O_BINARY
-#define O_BINARY 0
-#endif
 
 int main(int argc, char **argv) {
   if (argc < 2) {
